+++ packages/client/src/Call.ts
@@ -376,7 +376,7 @@ export class Call {
         const currentUserId = this.currentUserId;
         if (currentUserId && blockedUserIds.includes(currentUserId)) {
           this.logger('info', 'Leaving call because of being blocked');
-          await this.leave();
+          await this.leave({ reason: 'user blocked' });
         }
       }),
     );
@@ -463,7 +463,10 @@ export class Call {
   /**
    * Leave the call and stop the media streams that were published by the call.
    */
-  leave = async ({ reject = false }: CallLeaveOptions = {}) => {
+  leave = async ({
+    reject = false,
+    reason = 'user is leaving the call',
+  }: CallLeaveOptions = {}) => {
     const callingState = this.state.callingState;
     if (callingState === CallingState.LEFT) {
       throw new Error('Cannot leave call that has already been left.');
@@ -498,7 +501,7 @@ export class Call {
     this.publisher?.close();
     this.publisher = undefined;
 
-    this.sfuClient?.close();
+    this.sfuClient?.close(StreamSfuClient.NORMAL_CLOSURE, reason);
     this.sfuClient = undefined;
 
     this.dispatcher.offAll();
@@ -744,7 +747,8 @@ export class Call {
      * A closure which hides away the re-connection logic.
      */
     const reconnect = async (
-      strategy: 'full' | 'fast' | 'migrate' = 'full',
+      strategy: 'full' | 'fast' | 'migrate',
+      reason: string,
     ): Promise<void> => {
       const currentState = this.state.callingState;
       if (
@@ -781,7 +785,7 @@ export class Call {
       if (strategy === 'fast') {
         sfuClient.close(
           StreamSfuClient.ERROR_CONNECTION_BROKEN,
-          'js-client: attempting fast reconnect',
+          `attempting fast reconnect: ${reason}`,
         );
       } else if (strategy === 'full') {
         // in migration or recovery scenarios, we don't want to
@@ -801,7 +805,7 @@ export class Call {
         // clean up current connection
         sfuClient.close(
           StreamSfuClient.NORMAL_CLOSURE,
-          'js-client: attempting full reconnect',
+          `attempting full reconnect: ${reason}`,
         );
       }
       await this.join({
@@ -811,10 +815,7 @@ export class Call {
 
       // clean up previous connection
       if (strategy === 'migrate') {
-        sfuClient.close(
-          StreamSfuClient.NORMAL_CLOSURE,
-          'js-client: attempting migration',
-        );
+        sfuClient.close(StreamSfuClient.NORMAL_CLOSURE, 'attempting migration');
       }
 
       this.logger(
@@ -870,7 +871,7 @@ export class Call {
           'info',
           `[Migration]: Going away from SFU... Reason: ${GoAwayReason[reason]}`,
         );
-        reconnect('migrate').catch((err) => {
+        reconnect('migrate', GoAwayReason[reason]).catch((err) => {
           this.logger(
             'warn',
             `[Migration]: Failed to migrate to another SFU.`,
@@ -905,14 +906,16 @@ export class Call {
         if (this.reconnectAttempts < this.maxReconnectAttempts) {
           sfuClient.isFastReconnecting = this.reconnectAttempts === 0;
           const strategy = sfuClient.isFastReconnecting ? 'fast' : 'full';
-          reconnect(strategy).catch((err) => {
-            this.logger(
-              'error',
-              `[Rejoin]: ${strategy} rejoin failed for ${this.reconnectAttempts} times. Giving up.`,
-              err,
-            );
-            this.state.setCallingState(CallingState.RECONNECTING_FAILED);
-          });
+          reconnect(strategy, `SFU closed the WS with code: ${e.code}`).catch(
+            (err) => {
+              this.logger(
+                'error',
+                `[Rejoin]: ${strategy} rejoin failed for ${this.reconnectAttempts} times. Giving up.`,
+                err,
+              );
+              this.state.setCallingState(CallingState.RECONNECTING_FAILED);
+            },
+          );
         } else {
           this.logger(
             'error',
@@ -940,7 +943,10 @@ export class Call {
         do {
           try {
             sfuClient.isFastReconnecting = isFirstReconnectAttempt;
-            await reconnect(isFirstReconnectAttempt ? 'fast' : 'full');
+            await reconnect(
+              isFirstReconnectAttempt ? 'fast' : 'full',
+              'Network: online',
+            );
             return; // break the loop if rejoin is successful
           } catch (err) {
             this.logger(
@@ -1061,7 +1067,7 @@ export class Call {
           await this.publisher.restartIce();
         } else if (previousSfuClient?.isFastReconnecting) {
           // reconnection wasn't possible, so we need to do a full rejoin
-          return await reconnect('full').catch((err) => {
+          return await reconnect('full', 're-attempting').catch((err) => {
             this.logger(
               'error',
               `[Rejoin]: Rejoin failed forced full rejoin.`,
@@ -1129,7 +1135,7 @@ export class Call {
           `[Rejoin]: Rejoin ${this.reconnectAttempts} failed.`,
           err,
         );
-        await reconnect();
+        await reconnect('full', 'previous attempt failed');
         this.logger(
           'info',
           `[Rejoin]: Rejoin ${this.reconnectAttempts} successful!`,
@@ -1852,7 +1858,7 @@ export class Call {
 
         clearTimeout(this.dropTimeout);
         this.dropTimeout = setTimeout(() => {
-          this.leave().catch((err) => {
+          this.leave({ reason: 'ring: timeout' }).catch((err) => {
             this.logger('error', 'Failed to drop call', err);
           });
         }, timeoutInMs);