+++ packages/client/src/StreamVideoClient.ts
@@ -232,7 +232,7 @@ export class StreamVideoClient {
         // if `call.created` was received before `call.ring`.
         // In that case, we cleanup the already tracked call.
         const prevCall = this.writeableStateStore.findCall(call.type, call.id);
-        await prevCall?.leave();
+        await prevCall?.leave({ reason: 'cleaning-up in call.ring' });
         // we create a new call
         const theCall = new Call({
           streamClient: this.streamClient,
@@ -310,7 +310,7 @@ export class StreamVideoClient {
    * Creates a new call.
    *
    * @param type the type of the call.
-   * @param id the id of the call, if not provided a unique random value is used
+   * @param id the id of the call.
    */
   call = (type: string, id: string) => {
     return new Call({