+++ packages/client/src/devices/__tests__/InputMediaDeviceManagerState.test.ts
@@ -54,7 +54,7 @@ describe('InputMediaDeviceManagerState', () => {
       expect(permissionStatus.addEventListener).toHaveBeenCalled();
     });
 
-    it('should emit false when prompt is needed', async () => {
+    it('should emit true when prompt is needed', async () => {
       const permissionStatus: Partial<PermissionStatus> = {
         state: 'prompt',
         addEventListener: vi.fn(),
@@ -63,30 +63,6 @@ describe('InputMediaDeviceManagerState', () => {
       globalThis.navigator ??= {} as Navigator;
       // @ts-ignore - navigator is readonly, but we need to mock it
       globalThis.navigator.permissions = { query };
-
-      const hasPermission = await new Promise((resolve) => {
-        state.hasBrowserPermission$.subscribe((v) => resolve(v));
-      });
-      expect(hasPermission).toBe(false);
-      expect(query).toHaveBeenCalledWith({ name: 'camera' });
-      expect(permissionStatus.addEventListener).toHaveBeenCalled();
-    });
-
-    it('should emit true when prompt is needed in Safari', async () => {
-      const permissionStatus: Partial<PermissionStatus> = {
-        state: 'prompt',
-        addEventListener: vi.fn(),
-      };
-      const query = vi.fn(() => Promise.resolve(permissionStatus));
-      globalThis.navigator ??= { userAgent: 'safari' } as Navigator;
-      // @ts-ignore - navigator is readonly, but we need to mock it
-      globalThis.navigator.permissions = { query };
-
-      Object.defineProperty(globalThis.navigator, 'userAgent', {
-        get() {
-          return 'safari';
-        },
-      });
 
       const hasPermission = await new Promise((resolve) => {
         state.hasBrowserPermission$.subscribe((v) => resolve(v));