+++ packages/client/src/store/CallState.ts
@@ -98,6 +98,15 @@ export enum CallingState {
 }
 
 /**
+ * Returns the default egress object - when no egress data is available.
+ */
+const defaultEgress: EgressResponse = {
+  broadcasting: false,
+  hls: { playlist_url: '' },
+  rtmps: [],
+};
+
+/**
  * Holds the state of the current call.
  * @react You don't have to use this class directly, as we are exposing the state through Hooks.
  */
@@ -983,15 +992,15 @@ export class CallState {
   };
 
   private updateFromHLSBroadcastStopped = () => {
-    this.setCurrentValue(this.egressSubject, (egress) => ({
-      ...egress!,
+    this.setCurrentValue(this.egressSubject, (egress = defaultEgress) => ({
+      ...egress,
       broadcasting: false,
     }));
   };
 
   private updateFromHLSBroadcastingFailed = () => {
-    this.setCurrentValue(this.egressSubject, (egress) => ({
-      ...egress!,
+    this.setCurrentValue(this.egressSubject, (egress = defaultEgress) => ({
+      ...egress,
       broadcasting: false,
     }));
   };
@@ -999,11 +1008,11 @@ export class CallState {
   private updateFromHLSBroadcastStarted = (
     event: CallHLSBroadcastingStartedEvent,
   ) => {
-    this.setCurrentValue(this.egressSubject, (egress) => ({
-      ...egress!,
+    this.setCurrentValue(this.egressSubject, (egress = defaultEgress) => ({
+      ...egress,
       broadcasting: true,
       hls: {
-        ...egress!.hls,
+        ...egress.hls,
         playlist_url: event.hls_playlist_url,
       },
     }));